pipelines/rj_smas__disparo_pic/pyproject.toml
[project]
name = "rj-smas-disparo-pic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.29.0",
    "pandas>=2.3.3",
    "prefect>=3.4.9",
    "pyyaml>=6.0.3",
]
